feat: Update dashboard permissions [KT-471]
